Working version of DataGrid refactor
